refactor: restructure directory configuration
