+++ lib/core/exception/exception_handler.dart
@@ -5,8 +5,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:trip_app_nativeapp/core/exception/api_exception.dart';
 import 'package:trip_app_nativeapp/features/auth/data/repositories/firebase_auth_repository.dart';
 
+import '../../view/widgets/helpers/scaffold_messenger.dart';
 import '../constants/string.dart';
-import '../view/widgets/helpers/scaffold_messenger.dart';
 
 final exceptionHandler = Provider<ExceptionHandler>(ExceptionHandler.new);
