+++ lib/features/auth/controller/auth_controller.dart
@@ -1,12 +1,12 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:trip_app_nativeapp/core/view/widgets/loading.dart';
 import 'package:trip_app_nativeapp/features/auth/domain/interactor/auth_interactor.dart';
+import 'package:trip_app_nativeapp/view/widgets/loading.dart';
 
-import '../../../../core/enum/login_type.dart';
-import '../../../../core/exception/exception_handler.dart';
-import '../../../../core/view/widgets/helpers/scaffold_messenger.dart';
-import '../../data/repositories/firebase_auth_repository.dart';
+import '../../../core/enum/login_type.dart';
+import '../../../core/exception/exception_handler.dart';
+import '../../../view/widgets/helpers/scaffold_messenger.dart';
+import '../data/repositories/firebase_auth_repository.dart';
 
 final firebaseAuthUserProvider = StreamProvider<User?>(
   (ref) => ref.watch(firebaseAuthProvider).userChanges(),
