+++ lib/view/pages/constant_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../debug/logger.dart';
+import '../../core/debug/logger.dart';
 import '../widgets/loading.dart';
 
 /// MaterialApp.router の builder で返すことによって
