+++ lib/view/pages/loading_page.dart
